Move breadcrumbs back to cells.tsx
